core: minor speed improvements (made use of fadvise on input fd and TCP_CORK on output fd if multiple chunks to send)
- do TCP_CORK on client socket if we're about to send more than one chunk

- do fadvise to hint the kernel that we're about to sequencially read
the file from beginning to its end.
I'm not yet quite happy with the current solution. i need to benchmark
it a little more.

Finally, i've renamed composite_buffer_async_writer::impl to context
as this is what it is, and i'm actually going to abstract the visitor
implementations anyways, to provide different writer backends, so i'd
like to reduce confusion.