refactored from simple stat_service to fancy-featured fileinfo_service
the fileinfo object does not just contain the stat struct, but also
- cached etag
- cached Last-Modified value

and maybe more to follow.

This is the successor to stat_service, the only thing missing, is that
it's not (yet?) using a cache<> for storing the fileinfo object but a
real map (which doesn't free entries when becoming too full)
We trust on the swappiness of the underlying OS though ;)

Also, started rewriting the upstream x0d.conf to the upcoming
configuration file format.