made buffer API more ASIO friendly!!!
Consider a construct like this:

	// read
	char data[1024];
	std::size_t n = asio::read(socket, asio::buffer(data));

	// write
	asio::write(socket, asio::buffer(data, n));

then you're now able to do the following:

	// read
	x0::buffer data(1024);
	data.resize( asio::read(socket, data.asio_avail()) );

	// write
	asio::write(socket, data.asio_filled());

where the member functions
	- asio_filled() maps to the already filled space inside the x0::buffer object, and
	- asio_avail() maps to the yet available space from end() to (capacity() - size()).

Unlike the buffer class itself, buffer views only provide one asio
mapping as a view can't be, but indeed could be able to point to the
available space of the underlying buffer.

This change is required in order to allow writing the HTTP request
directly into a big x0::buffer blob, that can then just be referenced
by x0::buffer::view objects to point to request header keys/values of
interest.