fd_source: added support for pread() along with read()
depending on whether we want offset/count to be != -1
pread() is used and read() otherwise.

We do this in order to not have to call lseek and friends to seek inside
the stream, e.g. when sending multiple different chunks of a file.
However, we still need to support native read() aswell,
as the file descriptor might be an unseekable file descriptor
(like sockets/pipes), too.