big fat wip-commit with focus on integrating the filter mechanism to x0 core and (once again) moving out io/filter code into the x0/io directory.
this is certainly NOT compilable at all, that is, the compile runs
up to plugins/sendfile.cpp, which needs some love not only in adapting
API changes to the plugin but also in what exactly I want to have
asynchronousely handled.

content generation should for sure be asynchronous, which makes it a
little fun with multiple HTTP Ranges, but I'll take on that. :)