core: migrate from Asio to libev.
This is not final, however, if it prooves to be the better performance
with all respects, I'll keep it this way.
In case of a revert, I'll try to keep the core API as best as it is (to
only change the most important bits).
If I'll keep libev as event loop, I *SHOULD* provide some kind of Asio
integration support.

This is also not just a plain move from Asio to libev but also contains
some important tweaks, fixes and improvements which shouldn't get
reverted if I'll decide to switch back to Asio.

Although, do not forget those few TODO comments introduced in the code.