core: initial work on built-in SSL support using GnuTLS.
Sure, this implementation could receive some more love:
- make SSL-enabled listener more configurable
- implement SSL-session caching (which improves performance dramatically)
- find a way to let the SSL-code not be to intrusive into the core (cleaner design!)

This version currently is fully built into the core.
I'll need to find either a way to abstract the SSL code and core API in
a way to let it be a standalone-plugin or just throw the current ssl
plugin ideas off the board.