[http] improve HTTP/1.1 request content processing.
I had to change the way in processing HTTP message pipelining, but the
new strategy is better anyways.

However, it is a little bit strange, that I currently have to keep the
*whole* buffer of a connection in-memory. Just think about
a 4 GB file upload - that would all have to stay in ram - just to
preserve HTTP pipelining.

One possible (future) solution might be, to always crop the
buffer's size and parse-offset back to header_end-state when currently
processing a content chunk (TODO: brainstorming).