fastcgi: initial draft-plugin. *basically* working
There is still much room to improve, but this shall finally
replace the cgi plugin in favor of the cgi-fcgi wrapper
and the increasing difficulty to maintain the cgi plugin in
the future - when x0 is to become multi threaded, etc.

this plugin shall primary support connecting to fastcgi
applications via TCP/IP, though, I recomment the spawn-fcgi
to also support fastcgi apps not supporting this method.