added preliminary systemd socket-based activation support
this works by checking whether or not we have received a file descriptor
that matches the requested bind address + port criteria from the
configuration file (e.g. via 'listen' or 'ssl.listen' statement).

If systemd did not provide the requested socket (or no systemd is
running) then it is still tried to create this socket manually.