[http] HttpConnection: code cleanup wrt isOpen/isClosed/isAborted - the last survived for good.
A connection is always considered open unless nobody invoked abort() on it (done on I/O errors).
A close() on a connection effectivily *kills* the object, by deleting it, which
is why we don't need any isClosed() anyway.

Thus, isAborted() is the only thing that makes sense to keep as we now support
some kind of soft-invalid requests and connections, which is, user code may
still access it even if we're not connected to the remote end-point anymore
(read: the connection was aborted, e.g. due to I/O errors).