[http] introduce custom error handlers and helpers.
new flow symbols:
- function void error.handler(proc);
- property int req.status;
- function void req.rewrite(string path);

One interesting side-effect feature, is, that you can chain error
handlers as deep as you want.
That is, if you set another error handler within your error handler,
than this will be invoked in case of an error.
If not done so, but the error handler generated an error itself,
the default builtin error page generator is used.

handler main {
	error.handler my_error
	docroot '/var/www/localhost/htdocs'
	staticfile
}

handler my_error {
	# we inherit the req.docroot from the prior setting
	req.rewrite '/custom-error.html'
	staticfile
}