[plugins] proxy: fixes passing X-Forwarded-Proto and X-Forwarded-For
We should pass those request headers to the backend, as the backend
server is only intersted in the proto being passed by the most-left
frontend node (e.g. when you chain reverse proxy nodes), to properly
handle computation of fully qualified in-site URLs and
avoid endless redirection loops when enforcing SSL.

For X-Forwarded-For, the left-most value is also more interesting than
the inner chained reverse proxies, as this will be the one of the
real client, ideally.