sql: print error message to stderr on sql statement execution error
It might or might not be a good idea, as a library, to write
shamelessly to stderr, however, I might make this a requirement,
that fd 2 must be connected to some logging component (e.g. stderr,
syslog, logfile, etc)
GUI apps may still feel free to use the error handling methods.