[http] HttpRequest: fixes a behavioural bug related to accesslog plugin not showing proper bytes-transmitted value
The accesslog plugin is also to write how many bytes have been actually
been transmitted to the client.
For that to function the on-request-complete callback must be invoked
once the response object has been not just flushed but also fully
being transmitted to the client *and* all request/connection properties
have been updated accordingly,
most importantly the HttpConnection::bytesTransmitted_ field.

This new solution defers write-callbacks out of the writeSome() call,
so that writeSome gets its chance to update everything properly.

UNFORTUNATELY, I decided to use C++11 lambda expressions to actually
defer the callbacks, which is very handy to use and looks nice and clean
but impacts performance *by an unknown factor* when compared
to the last solution with just one level of indirection.