[plugins] HttpHealthMonitor: fixes a bug that caused (kind of) okay response bodies to take the backend down.
When a HTTP server replies with a HTTP/1.1 200, sets `Connection: close`
and does neither provide a Content-Length nor `Transfer-Encoding: chunked`
and the connection had been closed at the end of the well formed response,
then the monitor would have set the backend to Offline due to
an early connection close.

That was due to a bug in HttpMessageProcessor missing exactly that one
case and then us not indicating the end.