[plugins] fastcgi: fixes edge-case assertion failures.
The fastcgi bugfix (amongst the debugging improvements) has been
also applied to director's fastcgi backend implementation.

I need to get rid of the code duplication soon, ideally by
adding a director.pass() handler to on-the-fly pass
a request to a single static backend and thus deprecating fastcgi and
proxy plugins.