[base] Socket: backgrounding connect() on UNIX domain socket impossible
At least this is what the attempt resulted in.
connect() on a nonblocking AF_UNIX socket (that would block otherwise)
does *not* return EINPROGRESS (just like AF_INET sockets) but returns
EAGAIN.

I then cannot just watch on the socket file descriptor for
write-readiness and then invoke getsocketopt(... SO_ERROR...) and
check if the connect completed, in fact, the latter indicated
success but a read/write operation on the socket results into an EINVAL
and just trying to actually invoke connect() again doesn't result
into the expected behaviour either, that is, that the connect() would
succeed.