[http] HttpRequest: fixing confusion and implicit bugs of URL encoded/unencoded values in uri/path/query-string.
the query string must be url encoded while the path must be operated
URL decoded on. The classical URL decoded `uri` property of HttpRequest
was useless, but we added a useful `unparsedUri` instead.
The `path` property now unfortunately has to be a Buffer instead of
just a BufferRef however.

The method urldecode() is a bit of code duplication for now,
but we might add a static Url::decode() and Url::encode() helper
to aid this, as this method is also used in some plugin(s).

As the QUERY_STRING of fastcgi/proxy/director plugin is now URL encoded
we should now have eliminated the bug Sabri ran into as well.