[http] HttpRequest: greatly improve Request-URI parsing optimized directory traversal detection.
Directory traversal detection now does not rely on realpath() glibc
function anymore, which is a context switch beast, and thus, should
be avoided at all cost within the hot path.

So we now, while traversing through the Request-URI, we count
the directory-depth and store it within HttpRequest::directoryDepth_
that is then checked in HttpRequest::testDirectoryTraversal() to
bail out if this value became negative, which means, that the
Request-URI points to some file outside its document root.