[plugins] director: fixes #44
PHP-FPM sends out EndRequest-records that inform the frontend that the
given request has been completed (in success or not).
However, a `phpinfo();` does not auto-generate a Status: response header
to tell us the response status code, and that's why we defaulted to 503
as the backend did NOT generate content (but did indeed set some
response headers).

Now, upon EndRequest, we test the protocol status and conditionally
set the response status code, if not yet set already.