[base] added support for multi accept().
When compiled with -DWITH_MULTI_ACCEPT=ON, you can pass multi_accept
option to a listener (e.g. listen and ssl.listen) with the number
of maximum attempts to accept() an incoming connection (defaults to 1).
If the accept() syscall returns EAGAIN the multi-accept loop will
be left early.

This change should help performance on servers with very high number
of incoming connections per second.

WITH_MULTI_ACCEPT defaults to OFF.