[http] HttpConnection: fixes bug with opportunistic write and TCP_DEFER_ACCEPT enabled at the same time.
When TCP_DEFER_ACCEPT and opportunistic writes are enabled at compile
time, it may happen that the whole request processing can be done
in a non-blocking manner - regarding accept() + read() + write().

However, when this happens, the next request (e.g. when keep-alived)
has not been processed because nobody told the client socket to watch
for new input, which is now fixed by emulating the ev::READ event
within the start() call instead of directly and just calling readSome().