[base] TokenShaper: new API for generic "traffic shaping".
This implementation is based on how Linux HTB is working but adapted
to the needs of our requirements, being able to priorize requests
in the director load balancer plugin later.

This commit also introduces a new test using the Google Test Framework.
On Ubuntu/Debian you need to `apt-get install libgtest-dev` as well
as `cd /usr/src/gtest && cmake && make && cp libgtest*.so /usr/local/lib`
in order to make use of it here.