Executor API gains reference counts.
In the Executor API, the reference count is incremented on each
task yet to be executed and decremented once executed.

The reference count is the number of tasks still pending which helps
runLoop() to determine wheather to continue iterating over
runLoopOnce() or to return to its caller.

Sometimes you need to manually decrement the reference count in order
to explicitely ignore a pending task in the loop.

However, you must increment the reference manually as soon
as your task has either been invoked or is to be cancelled to not
cause bugs in counting.